changed the criteria for the best weight file
changed the criteria for the best weight file from loss to mAP

I trained the model on my custom dataset. But I failed to get a good results when I load the weight file that has the lowest loss on test dataset. 

I thought that the loss used in YOLO is not proper criteria for detection performance. So I changed the criteria from loss to mAP.

what do you think of this?