Modified build_targets for test
When I trained and tested custom model with custom dataset, I had got NaN loss error during the test.

That problem is caused by reject flag that used to help train model in training process. After I removed that flag and condition, that problem is solved.

It seems like to cause error if we do not have enough anchors to compatible with the size of bounding box in test dataset.

So I added the condition to neglect reject flag during the test.

What do you think of this?