Update LICENSE to AGPL-3.0
This pull request updates the license of the YOLOv3 project from GNU General Public License v3.0 (GPL-3.0) to GNU Affero General Public License v3.0 (AGPL-3.0).

We at Ultralytics have decided to make this change in order to better protect our intellectual property and ensure that any modifications made to the YOLOv3 source code will be shared back with the community when used over a network.

AGPL-3.0 is very similar to GPL-3.0, but with an additional clause to address the use of software over a network. This change ensures that if someone modifies YOLOv3 and provides it as a service over a network (e.g., through a web application or API), they must also make the source code of their modified version available to users of the service.

This update includes the following changes:

Replace the LICENSE file with the AGPL-3.0 license text
Update the license reference in the README.md file
Update the license headers in source code files
We believe that this change will promote a more collaborative environment and help drive further innovation within the YOLOv3 community.